merge master

+++ upload/system/library/session/file.php
@@ -1,6 +1,10 @@
 <?php
 namespace Opencart\System\Library\Session;
 class File {
+	public function __construct($registry) {
+		$this->config = $registry->get('config');
+	}
+
 	public function read($session_id) {
 		$file = DIR_SESSION . 'sess_' . basename($session_id);
 
@@ -53,28 +57,14 @@ class File {
 		}
 	}
 
-	public function __destruct() {
-		$gc_divisor = (int)ini_get('session.gc_divisor');
-
-		if ($gc_divisor) {
-			$gc_divisor = $gc_divisor;
-		} else {
-			$gc_divisor = 1;
-		}
-
-		if (ini_get('session.gc_probability')) {
-			$gc_probability = ini_get('session.gc_probability');
-		} else {
-			$gc_probability = 1;
-		}
-
-		if (mt_rand() / mt_getrandmax() < $gc_probability / $gc_divisor) {
-			$expire = time() - ini_get('session.gc_maxlifetime');
+	public function gc() {
+		if (round(rand(1, $this->config->get('session_divisor') / $this->config->get('session_probability'))) == 1) {
+			$expire = time() - $this->config->get('session_expire');
 
 			$files = glob(DIR_SESSION . 'sess_*');
 
 			foreach ($files as $file) {
-				if (filemtime($file) < $expire) {
+				if (is_file($file) && filemtime($file) > $expire) {
 					unlink($file);
 				}
 			}
